Remove (more or less) empty files
