--- libs/unity/library/Runtime/Scripts/Media/IAudioSource.cs
@@ -1,22 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-using System;
-using UnityEngine.Events;
-
-namespace Microsoft.MixedReality.WebRTC.Unity
-{
-    /// <summary>
-    /// Unity event corresponding to a new audio stream being started.
-    /// </summary>
-    [Serializable]
-    public class AudioStreamStartedEvent : UnityEvent<IAudioSource>
-    { };
-
-    /// <summary>
-    /// Unity event corresponding to an on-going audio stream being stopped.
-    /// </summary>
-    [Serializable]
-    public class AudioStreamStoppedEvent : UnityEvent<IAudioSource>
-    { };
-}
