--- libs/unity/library/Runtime/Scripts/Media/IAudioSource.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e1ae9d533dc442a429f9db2a3fdf51a1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {fileID: 2800000, guid: f845cb707dda5c84594956f49655100f, type: 3}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
