--- libs/unity/library/Runtime/Scripts/Media/IVideoSource.cs
@@ -1,22 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-using System;
-using UnityEngine.Events;
-
-namespace Microsoft.MixedReality.WebRTC.Unity
-{
-    /// <summary>
-    /// Unity event corresponding to a new video stream being started.
-    /// </summary>
-    [Serializable]
-    public class VideoStreamStartedEvent : UnityEvent<IVideoSource>
-    { };
-
-    /// <summary>
-    /// Unity event corresponding to an on-going video stream being stopped.
-    /// </summary>
-    [Serializable]
-    public class VideoStreamStoppedEvent : UnityEvent<IVideoSource>
-    { };
-}
