--- libs/unity/library/Runtime/Scripts/Media/IVideoSource.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2f626969ae477c744813ca1727b3d63d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {fileID: 2800000, guid: f845cb707dda5c84594956f49655100f, type: 3}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
